pyproject.toml
[project]
name = "qbpm"
version = "1.0-rc2"
description = "qutebrowser profile manager"
license = { file = "LICENSE" }
readme = "README.md"
authors = [{ name = "Peter Rice", email = "peter@peterrice.xyz" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Typing :: Typed",
]
requires-python = ">= 3.9"
dependencies = ["pyxdg", "click"]

[project.urls]
repository = "https://github.com/pvsr/qbpm"

[project.scripts]
qbpm = "qbpm.main:main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["qbpm"]

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "xdg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.ruff]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "N",
    "ANN",
    "B",
    "A",
    "C4",
    "PT",
    "SIM",
    "ARG",
    "PTH",
    "PL",
    "RUF"
]
target-version = "py39"
ignore = [
  "E501", # line too long
  "ANN101", # annotate self
  "ANN102", # annotate cls
  "ANN401", # ban Any
]

[tool.ruff.per-file-ignores]
"tests/test_main.py" = [ "S101", "ANN201"]
"tests/test_profiles.py" = [ "S101", "ANN201"]
